🚧 Apply JSDocs to api functions
