+++ src/api/options.js
@@ -1,6 +1,11 @@
 import { collection, doc, getDoc } from 'firebase/firestore'
 import { db } from '../../firebase.config'
 
+/**
+ * 데이터베이스에서 옵션을 불러온다.
+ *
+ * @returns {Object.<string, any>[]?} options
+ */
 async function getOptions() {
   try {
     const collectionRef = collection(db, 'sikmechu')
